### 2019-06-30 #1537 @whedon
[PRE REVIEW]: bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework
**Submitting author:** @DominiqueMakowski (<a href="http://orcid.org/0000-0001-5375-9967">Dominique Makowski</a>)
**Repository:** <a href="https://github.com/easystats/bayestestR" target ="_blank">https://github.com/easystats/bayestestR</a>
**Version:** v0.2.3
