### 2019-07-01 4880ddf
updated function docs
#177
