+++ R/bayesfactor_parameters.R
@@ -1,17 +1,20 @@
 #' Savage-Dickey density ratio Bayes Factor (BF)
 #'
-#' This method computes the ratio between the density of a single value (typically the null)
-#' of two distributions. When the compared distributions are the posterior and the prior distributions,
-#' this results is an approximation of a Bayes factor comparing the model against a model in which
-#' the parameter of choice is restricted to the point null.
+#' This method computes Bayes factors against the null (either a point or an interval),
+#' bases on prior and posterior samples of a single parameter. When the null is a point,
+#' a Savage-Dickey density ratio is computed; when the null is an interval, the Bayes
+#' factor is computed by comparing the prior and posterior odds of the parameter falling
+#' within or outside the null.
 #' \cr \cr
-#' See also \href{https://easystats.github.io/bayestestR/articles/bayes_factors.html}{the Bayes factors vignette}.
+#' For more info, see \href{https://easystats.github.io/bayestestR/articles/bayes_factors.html}{the Bayes factors vignette}.
 #'
-#' @param posterior A numerical vector, \code{stanreg} / \code{brmsfit} object, \code{emmGrid} or a data frame - representing a posterior distribution(s) from (see Details).
+#' @param posterior A numerical vector, \code{stanreg} / \code{brmsfit} object, \code{emmGrid}
+#' or a data frame - representing a posterior distribution(s) from (see Details).
 #' @param prior An object representing a prior distribution (see Details).
 #' @param direction Test type (see details). One of \code{0}, \code{"two-sided"} (default, two tailed),
 #' \code{-1}, \code{"left"} (left tailed) or \code{1}, \code{"right"} (right tailed).
-#' @param null Value of the null model. If can be a scaler (for point-null model) or a a range (for a interval-null model). Defaults to \code{0} for a "classic" point-null test.
+#' @param null Value of the null, either a scaler (for point-null) or a a range
+#' (for a interval-null).
 #' @inheritParams hdi
 #'
 #' @return A data frame containing the Bayes factor representing evidence \emph{against} the (point) null effect model.
@@ -20,7 +23,8 @@
 #' When \code{posterior} is a model (\code{stanreg}, \code{brmsfit}), posterior and prior samples are
 #' extracted for each parameter, and Savage-Dickey Bayes factors are computed for each parameter.
 #'
-#' \strong{NOTE:} For \code{brmsfit} models, the model must have been fitted with \emph{custom (non-default)} priors. See example below.
+#' \strong{NOTE:} For \code{brmsfit} models, the model must have been fitted with \emph{custom (non-default)}
+#' priors. See example below.
 #'
 #' \subsection{Setting the correct \code{prior}}{
 #' It is important to provide the correct \code{prior} for meaningful results.
@@ -30,10 +34,12 @@
 #'   \item When \code{posterior} is a \code{stanreg} or \code{brmsfit} model, there is no need to specify \code{prior}, as prior samples are drawn internally.
 #'   \item When \code{posterior} is a \code{data.frame}, \code{prior} should also be a \code{data.frame}, with matching column order.
 #' }}
-#' \subsection{One-sided & Intervall-Null Tests (setting an order restriction)}{
-#' ## EDIT ##
-#' One sided tests (controlled by \code{direction}) are conducted by setting an order restriction on
-#' the prior and posterior distributions (\cite{Morey & Wagenmakers, 2013}).
+#' \subsection{One-sided Tests (setting an order restriction)}{
+#' One sided tests (controlled by \code{direction}) are conducted by restricting the prior and
[…]
