### 2019-07-01 811dae2
updated vignette
#177
