+++ vignettes/bayes_factors.Rmd
@@ -27,10 +27,19 @@ This vignette can be referred to by citing the package:
 
 ```{r setup, include=FALSE}
 library(knitr)
+library(rstanarm)
+library(bayestestR)
+library(ggplot2)
+library(see)
+library(emmeans)
+# library(brms)
+library(lme4)
+library(BayesFactor)
+
 options(knitr.kable.NA = '')
-knitr::opts_chunk$set(echo = TRUE)
-knitr::opts_chunk$set(comment = ">")
-ggplot2::theme_set(see::theme_modern())
+opts_chunk$set(echo = TRUE)
+opts_chunk$set(comment = ">")
+theme_set(see::theme_modern())
 options(digits = 2)
 set.seed(5)
 ```
@@ -41,14 +50,15 @@ The adoption of the Bayesian framework for applied statistics, especially in soc
 
 Having said that, here's an introduction to Bayes factors :)
 
+
 # Bayes Factors
 
-**Bayes factors (BFs) are indices of *relative* evidence of one "model" over another**, which are used in Bayesian inference as alternatives to classical (frequentist) hypothesis testing indices (such as $p-values$). In the Bayesian framework, a Bayes factor can also be thought of as the quantity by which some prior beliefs about the relative odds of two models are updated in light of the observed data.
+**Bayes factors (BFs) are indices of *relative* evidence of one "model" over another**, which are used in Bayesian inference as alternatives to classical (frequentist) hypothesis testing indices (such as $p-values$).
 
 According to Bayes' theorem:
 
 $$
-P(M|D) = \frac{P(D|M)P(M)}{P(D)}
+P(M|D) = \frac{P(D|M)\times P(M)}{P(D)}
 $$
 
 Then by comparing two models, we get:
@@ -60,15 +70,207 @@ Where the middle term is the Bayes factor:
 $$
 BF_{12}=\frac{P(D|M_1)}{P(D|M_2)}
 $$
-These are usually computed as the ratio of marginal likelihoods of two competing hypotheses / models, but as we can see from the equation above, they can also be computed by dividing the posterior odds by the prior odds. Importantly, **Bayes factors cover a wide range of indices and applications**, and come in different flavors.
+Importantly, **Bayes factors cover a wide range of indices and applications**, and come in different flavors. For example, in the Bayesian framework, a Bayes factor can also be seen as the quantity by which some prior beliefs about the relative odds of two models are updated in light of the observed data - and in fact one way to compute Bayes factors is to divide posterior odds by prior odds, as we will soon demonstrate.
 
-## Comparing models
+## Bayes Factors for a Single Parameter
 
-Generally speaking, Bayes factors are used to compare models and answer the question:
+A ***Bayes factor for a single parameter*** can be used to answer the question:
 
-> **Under which model is the the observed data more probable?**
+> **Given the observed data, is the null more, or less likely?**
 
-In other words, which model is more likely to have produced the observed data? This is usually done by computing the marginal likelihoods of two models. In such a case, the Bayes factor is a measure of the relative evidence of one of the compared models over the other.
+Let's use the Students' Sleep data set (`sleep`), and try and answer the question: ***given the observed data, is it more or less likely that the drug (the variable `group`) has no effect on the numbers of hours of extra sleep (variable `extra`)?***
+
+```{r sleep_boxplot, echo=FALSE, message=FALSE, warning=FALSE}
+ggplot(sleep, aes(x = group, y = extra, fill= group)) +
+  geom_boxplot() +
+  theme_classic()
+```
+
[…]
