### 2019-07-01 8175ce2
sensitivity to priors
close #189
