+++ NAMESPACE
@@ -131,6 +131,7 @@ S3method(rope_range,brmsfit)
 S3method(rope_range,lm)
 S3method(rope_range,merMod)
 S3method(rope_range,stanreg)
+S3method(sensitivity_to_prior,stanreg)
 S3method(update,bayesfactor_models)
 export(area_under_curve)
 export(auc)
@@ -174,6 +175,7 @@ export(probability_at)
 export(rnorm_perfect)
 export(rope)
 export(rope_range)
+export(sensitivity_to_prior)
 importFrom(insight,find_algorithm)
 importFrom(insight,find_formula)
 importFrom(insight,find_parameters)
