### 2019-07-01 #1537 @tjmahr
I can review it. (I have done one review in the past for ROpenSci.)
