### 2019-07-01 #1537 @paul-buerkner
It is a rather short article, right? I think I should be able to review it. Are there any reviewing guidelines?
