+++ NAMESPACE
@@ -1,13 +1,11 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(as.data.frame,density)
-S3method(as.double,bayesfactor_savagedickey)
 S3method(as.double,map_estimate)
 S3method(as.double,p_direction)
 S3method(as.double,p_map)
 S3method(as.double,p_rope)
 S3method(as.double,rope)
-S3method(as.numeric,bayesfactor_savagedickey)
 S3method(as.numeric,map_estimate)
 S3method(as.numeric,p_direction)
 S3method(as.numeric,p_map)
