### 2019-07-01 a172f05
allow intervals
