+++ vignettes/bayes_factors.Rmd
@@ -82,7 +82,13 @@ A ***Bayes factor for a single parameter*** can be used to answer the question:
 
 > **Given the observed data, is the null hypothesis of an absence effect (parameter = 0) more, or less likely?**
 
-Let's use the Students' Sleep data set (`sleep`), and try and answer the question: *given the observed data, how likely it is that the **drug** (the variable `group`) **has no effect** on the numbers of hours of extra sleep (variable `extra`)?*
+
+```{r echo=FALSE, fig.cap="Bayesian analysis of the Students' (1908) Sleep data set.", fig.align='center', out.width="80%"}
+knitr::include_graphics("https://github.com/easystats/easystats/raw/master/man/figures/bayestestR/deathsticks.png")
+```
+
+
+Let's use the Students' (1908) Sleep data set (`sleep`), in which **people took some drug** and where the researchers measured the number of hours of **extra hours of sleep** that they slept after. We will try answering the following question: *given the observed data, how likely it is that the **drug** (the variable `group`) **has no effect** on the numbers of hours of **extra sleep** (variable `extra`)?*
 
 ```{r sleep_boxplot, echo=FALSE, message=FALSE, warning=FALSE}
 ggplot(sleep, aes(x = group, y = extra, fill= group)) +
@@ -92,9 +98,10 @@ ggplot(sleep, aes(x = group, y = extra, fill= group)) +
 
 The **bloxplot** suggests that the 2nd group has a higher number of hours of extra sleep. By how much? Let's fit a simple [Bayesian linear model](https://easystats.github.io/bayestestR/articles/example1_GLM.html).
 
-### Null-Interval Bayes Factors
+### Testing against a Null-*Region*
 
-We can do this by setting an null interval, such that an effect within the interval would be practically equivalent to the drug having no effect at all. We can then compute the prior probability of the drug's effect falling *within the null-interval*, and the prior probability of the drug's effect falling *outside the null-interval* to get out prior odds. Say any effect smaller than half an hour of extra sleep is practically equivalent to being no effect at all, out prior odds would be:
+
+One way ofWe can do this by setting an null interval, such that an effect within the interval would be practically equivalent to the drug having no effect at all. We can then compute the prior probability of the drug's effect falling *within the null-interval*, and the prior probability of the drug's effect falling *outside the null-interval* to get out prior odds. Say any effect smaller than half an hour of extra sleep is practically equivalent to being no effect at all, out prior odds would be:
 
 $$
 \frac
