### 2019-07-01 #1537 @strengejacke
@paul-buerkner

Guidelines are described here: https://joss.readthedocs.io/en/latest/review_criteria.html
The review will be in a GitHub repo, where reviewers have this checklist:
https://joss.readthedocs.io/en/latest/review_checklist.html

You can glance at the compiled PDF here:
https://github.com/openjournals/joss-papers/blob/joss.01537/joss.01537/10.21105.joss.01537.pdf

But it's not only the review of the paper, but also of the GitHub-repository, as you have a checklist for both the paper and the repo.
