+++ vignettes/bayes_factors.Rmd
@@ -99,15 +99,15 @@ The **bloxplot** suggests that the 2nd group has a higher number of hours of ext
 ### Testing against a Null-*Region*
 
 
-One way of operationlizing the null-hypothesis is by setting a null region (for example the `[-0.1, 0.1]` interval), such that an effect that falls within this interval would be practically equivalent to the the null. In our case, that means defining a region where we would consider the drug having no effect at all. We can then compute the prior probability of the drug's effect falling *within the null-interval*, and the prior probability of the drug's effect falling *outside the null-interval* to get out prior odds. Say any effect smaller than half an hour of extra sleep is practically equivalent to being no effect at all, out prior odds would be:
+One way of operationlizing the null-hypothesis is by setting a null region (for example the `[-0.1, 0.1]` interval), such that an effect that falls within this interval would be practically equivalent to the the null. In our case, that means defining a region where we would consider the drug having no effect at all. We can then compute the prior probability of the drug's effect falling *within this null-region*, and the prior probability of the drug's effect falling *outside the null-region* to get out prior odds. Say any effect smaller than half an hour of extra sleep is practically equivalent to being no effect at all, out prior odds would be:
 
 $$
 \frac
-{P(b_{drug} \in [-\frac{1}{2},\frac{1}{2}])}
-{P(b_{drug} \notin [-\frac{1}{2},\frac{1}{2}])}
+{P(b_{drug} \in [-0.5, 0.5])}
+{P(b_{drug} \notin [-0.5, 0.5])}
 $$
 
-If we set our prior to have a normal distribution centered at 0 hours with a scale of 2.5 hours, or prior would look like this:
+If we set our prior to have a normal distribution centered at 0 hours with a scale (an SD) of 2.5 hours, or prior would look like this:
 
 ```{r, echo=FALSE}
 null <- c(-.5,.5)
@@ -142,12 +142,12 @@ library(rstanarm)
 model <- stan_glm(extra ~ group, data = sleep)
 ```
 
-By looking at the posterior distribution, can now compute the posterior probability of the drug's effect falling *within the null-interval*, and the posterior probability of the drug's effect falling *outside the null-interval* to get out posterior odds:
+By looking at the posterior distribution, can now compute the posterior probability of the drug's effect falling *within the null-region*, and the posterior probability of the drug's effect falling *outside the null-region* to get out posterior odds:
 
 $$
 \frac
-{P(b_{drug} \in [-\frac{1}{2},\frac{1}{2}] | Data)}
-{P(b_{drug} \notin [-\frac{1}{2},\frac{1}{2}] | Data)}
+{P(b_{drug} \in [-0.5,0.5] | Data)}
+{P(b_{drug} \notin [-0.5,0.5] | Data)}
 $$
 
 ```{r rstanarm_fit, echo=FALSE, message=FALSE, warning=FALSE}
@@ -188,7 +188,7 @@ BF <- My_first_BF$BF[2]
 post_odds <- prior_odds * BF
 ```
 
-We can see that the center of the posterior distribution has shifted away from 0 (to ~1.6). Likewise, the posterior odds are `r scales::number(post_odds, accuracy = 0.1)` - which seems to favor the effect being non-null, but... does this mean the data support the alternative more than the null? Hard to say, since even before the data were observed, the priors favored the alternative - so we need to take our priors into account here!
+We can see that the centre of the posterior distribution has shifted away from 0 (to ~1.6). Likewise, the posterior odds are `r scales::number(post_odds, accuracy = 0.1)` - which seems to favour **the effect being non-null**, but... *does this mean the data support the alternative more than the null?* Hard to say, since even before the data were observed, the priors already favoured the alternative - so we need to take our priors into account here!
 
[…]
