### 2019-07-02 #1541 @whedon
[REVIEW]: bayestestR: Describing Effects and their Uncertainty, Existence and Significance within the Bayesian Framework
**Submitting author:** @DominiqueMakowski (<a href="http://orcid.org/0000-0001-5375-9967">Dominique Makowski</a>)
**Repository:** <a href="https://github.com/easystats/bayestestR" target ="_blank">https://github.com/easystats/bayestestR</a>
**Version:** 0.2.5
