### 2019-07-04 bb02171
APA style to all vignettes
