### 2019-07-08 #1541 @paul-buerkner
I have just finished my review and have very few minor comments.

- All the citations in the software paper should list a DOI (if they have one) as per reviewer checklist above.
- The definition of the maximum a-posteriori value as "the most probable value" is not entirely correct for continuous parameters. Instead it is the value with the highest density (which still have probability zero for continuous parameters).
- you seem to use the abbreviation MAP both for the maximum a posteriori and the maximum a-priori value. This will likely confuse readers.
