### 2019-07-15 beea84f
bayesfactor_parameters.bayesfactor_models
error
