+++ NAMESPACE
@@ -16,6 +16,7 @@ S3method(bayesfactor_models,BFBayesFactor)
 S3method(bayesfactor_models,brmsfit)
 S3method(bayesfactor_models,default)
 S3method(bayesfactor_models,stanreg)
+S3method(bayesfactor_parameters,bayesfactor_models)
 S3method(bayesfactor_parameters,brmsfit)
 S3method(bayesfactor_parameters,data.frame)
 S3method(bayesfactor_parameters,emmGrid)
