### 2019-07-15 50b5b31
tests for bayesfactor_restricted
