### 2019-07-15 aa76ef5
updated various BF docs
