+++ vignettes/bayes_factors.Rmd
@@ -63,7 +63,6 @@ According to Bayes' theorem:
 $$
 P(M|D) = \frac{P(D|M)\times P(M)}{P(D)}
 $$
-
 Then by comparing two models, we get:
 
 $$
@@ -75,14 +74,13 @@ BF_{12}=\frac{P(D|M_1)}{P(D|M_2)}
 $$
 Thus, Bayes factors can be seen either as a ratio quantifying ***the relative likelihood of two models in light of some observed data*** as they can be computed by comparing marginal likelihoods, or as ***the degree by which some prior beliefs about the relative odds of two models are to be updated*** as they can be computed by dividing posterior odds by prior odds, as we will soon demonstrate.
 
-
-Critically, Bayes factors could be regrouped by two different applications: **testing single parameters (coefficients) within a model** and **comparing statistical models themselves**.
+Here we provide functions for computing Bayes factors in two different applications: **testing single parameters (coefficients) within a model** and **comparing statistical models themselves**.
 
 ## Testing Models' Parameters with Bayes Factors
 
 A ***Bayes factor for a single parameter*** can be used to answer the question:
 
-> **Given the observed data, is the null hypothesis of an absence effect more, or less likely?**
+> **Given the observed data, is the null hypothesis of an absence of an effect more, or less likely?**
 
 
 ```{r echo=FALSE, fig.cap="Bayesian analysis of the Students' (1908) Sleep data set.", fig.align='center', out.width="80%"}
@@ -90,7 +88,7 @@ knitr::include_graphics("https://github.com/easystats/easystats/raw/master/man/f
 ```
 
 
-Let's use the Students' (1908) Sleep data set (`sleep`), in which **people took some drug** and where the researchers measured the **extra hours of sleep** that they slept after. We will try answering the following question: *given the observed data, how likely it is that the **drug** (the variable `group`) **has no effect** on the numbers of hours of **extra sleep** (variable `extra`)?*
+Let's use the Students' (1908) Sleep data set (`data("sleep")`), in which **people took some drug** and where the researchers measured the **extra hours of sleep** that they slept afterwards. We will try answering the following question: *given the observed data, how likely it is that the **drug** (the variable `group`) **has no effect** on the numbers of hours of **extra sleep** (variable `extra`)?*
 
 ```{r sleep_boxplot, echo=FALSE, message=FALSE, warning=FALSE}
 ggplot(sleep, aes(x = group, y = extra, fill= group)) +
@@ -103,7 +101,7 @@ The **bloxplot** suggests that the 2nd group has a higher number of hours of ext
 ### Testing against a Null-*Region*
 
 
-One way of operationlizing the null-hypothesis is by setting a null region (for example the `[-0.1, 0.1]` interval), such that an effect that falls within this interval would be practically equivalent to the the null. In our case, that means defining a region where we would consider the drug having no effect at all. We can then compute the prior probability of the drug's effect falling *within this null-region*, and the prior probability of the drug's effect falling *outside the null-region* to get out prior odds. Say any effect smaller than half an hour of extra sleep is practically equivalent to being no effect at all, out prior odds would be:
[…]
