+++ R/bayesfactor.R
@@ -46,15 +46,15 @@
 #' @export
 bayesfactor <-
   function(...,
-           prior = NULL,
-           direction = "two-sided",
-           null = 0,
-           hypothesis = NULL,
-           effects = c("fixed", "random", "all"),
-           verbose = TRUE,
-           denominator = 1,
-           match_models = FALSE,
-           prior_odds = NULL) {
+             prior = NULL,
+             direction = "two-sided",
+             null = 0,
+             hypothesis = NULL,
+             effects = c("fixed", "random", "all"),
+             verbose = TRUE,
+             denominator = 1,
+             match_models = FALSE,
+             prior_odds = NULL) {
     mods <- list(...)
 
     if (length(mods) > 1) {
@@ -69,9 +69,10 @@ bayesfactor <-
       }
     } else if (!is.null(hypothesis)) {
       bayesfactor_restricted(...,
-                             prior = prior,
-                             verbose = verbose,
-                             effects = effects)
+        prior = prior,
+        verbose = verbose,
+        effects = effects
+      )
     } else {
       bayesfactor_parameters(
         ...,
