+++ man/sensitivity_to_prior.Rd
@@ -37,7 +37,6 @@ sensitivity_to_prior(model, index = c("Median", "MAP"))
 library(brms)
 model <- brms::brm(mpg ~ wt + cyl, data = mtcars)
 # sensitivity_to_prior(model)
-
 }
 }
 \seealso{

[…]
