+++ man/bayesfactor_parameters.Rd
@@ -2,8 +2,6 @@
 % Please edit documentation in R/bayesfactor_parameters.R
 \name{bayesfactor_parameters}
 \alias{bayesfactor_parameters}
-\alias{bayesfactor_savagedickey}
-\alias{bayesfactor_interval}
 \alias{bayesfactor_parameters.numeric}
 \alias{bayesfactor_parameters.stanreg}
 \alias{bayesfactor_parameters.brmsfit}
@@ -13,12 +11,6 @@
 \usage{
 bayesfactor_parameters(posterior, prior = NULL,
   direction = "two-sided", null = 0, verbose = TRUE, ...)
-
-bayesfactor_savagedickey(posterior, prior = NULL,
-  direction = "two-sided", null = 0, verbose = TRUE, ...)
-
-bayesfactor_interval(posterior, prior = NULL, direction = "two-sided",
-  null = rope_range(posterior), verbose = TRUE, ...)
 
 \method{bayesfactor_parameters}{numeric}(posterior, prior = NULL,
   direction = "two-sided", null = 0, verbose = TRUE, ...)
