+++ NAMESPACE
@@ -141,7 +141,6 @@ export(area_under_curve)
 export(auc)
 export(bayesfactor)
 export(bayesfactor_inclusion)
-export(bayesfactor_interval)
 export(bayesfactor_models)
 export(bayesfactor_parameters)
 export(bayesfactor_restricted)
