### 2019-07-15 3086e6d
fix pd_to_p, continue documentationg BF SD and swap BFSD appearances
