+++ DESCRIPTION
@@ -22,7 +22,11 @@ Authors@R: c(
 		"Wilson", 
 		role = c("aut"), 
 		email = "michael.d.wilson@curtin.edu.au", 
-		comment = c(ORCID = "0000-0003-4143-7308"))
+		comment = c(ORCID = "0000-0003-4143-7308")),
+	person("Paul-Christian", 
+		"Bürkner", 
+		role = c("rev"), 
+		email = "paul.buerkner@gmail.com"))
 	)
 Maintainer: Dominique Makowski <dom.makowski@gmail.com>
 URL: https://github.com/easystats/bayestestR
