### 2019-07-15 3de5a8f
Replace Maximum A Priori by Maximum A Posteriori
