### 2019-07-19 #1541 @DominiqueMakowski
Dear @paul-buerkner, thanks a lot for your comments! We addressed them in [this PR](https://github.com/easystats/bayestestR/pull/197#issue-297553261):

> Reviewer 1 (@paul-buerkner)
>
> * [x]  **All the citations in the software paper should list a DOI (if they have one) as per reviewer checklist above.**
> * Added DOIs for all refs but the following (none was found):
>
>   * `see` package ([here](https://github.com/easystats/see/issues/36))
>   * rstanarm package
>   * BayesFactor package
>   * Mill's "Objective Bayesian Precise Hypothesis Testing"
>   * Multiple Comparisons with BayesFactor, Part 2 (Morey's blog, 2015)
>   * Practical bayesian optimization of machine learning algorithms (Snoek's proceedings, 2012)
>   * Jeffrey's Theory of Probability book
> * [x]  **The definition of the maximum a-posteriori value as "the most probable value" is not entirely correct for continuous parameters. Instead it is the value with the highest density (which still have probability zero for continuous parameters).**
> * We changed its definition to the following:
>
> "find the **Highest Maximum A Posteriori (MAP)** estimate of a posterior, _i.e._, the value associated with the highest probability density (the "peak" of the posterior distribution). In other words, it is an estimation of the _mode_ for continuous parameters."
>
> * [x]  **you seem to use the abbreviation MAP both for the maximum a posteriori and the maximum a-priori value. This will likely confuse readers.**
> * This was likely an error and was addressed by replacing instances of the latter by the former (the maximum a posteriori).

Please note that there are still some references for which we did not find a DOI: we continue our search in the meantime. We hope you will find the revised version satisfying ☺️
