### 2019-07-21 809ea8d
Update example3.Rmd
