### 2019-07-21 b09498b
Update describe_posterior.R
