+++ R/describe_posterior.R
@@ -4,8 +4,8 @@
 #'
 #' @param posteriors A vector, dataframe or model of posterior draws.
 #' @param ci_method The type of index used for Credible Interval. Can be
-#'   \code{"hdi"} (default, see \code{\link{hdi}}) or \code{"quantile"}
-#'   (see \code{\link{ci}}).
+#'   \code{"HDI"} (default, see \code{\link{hdi}}) or \code{"ETI"}
+#'   (see \code{\link{eti}}).
 #' @param test The indices of effect existence to compute. Character (vector) or
 #'   list with one or more of these options: \code{"p_direction"} (or \code{"pd"}),
 #'   \code{"rope"}, \code{"p_map"}, \code{"equivalence_test"} (or \code{"equitest"}),
