### 2019-07-21 98b740e
minor fix
https://github.com/easystats/estimate/issues/14
