+++ vignettes/example2.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Example 2: Confirmation of Bayesian skills"
+title: "2. Confirmation of Bayesian skills"
 output: 
   github_document:
     toc: true
@@ -38,9 +38,9 @@ set.seed(333)
 ```
 
 
-Now that [**describing and understanding posterior distributions**](https://easystats.github.io/bayestestR/articles/example1.html) of linear regressions has no secrets for you, let's go back and study some simpler models: **correlations** and ***t*-tests**.
+Now that [**describing and understanding posterior distributions**](https://easystats.github.io/bayestestR/articles/example1.html) of linear regressions has no secrets for you, we will take one step back and study some simpler models: **correlations** and ***t*-tests**.
 
-But before we do that, let us take a moment to remind ourselves and appreciate the fact that **all basic statistical pocedures** such as correlations, *t*-tests, ANOVAs or Chisquare tests ***are* linear regressions** (we strongly recommend  [this excellent demonstration](https://lindeloev.github.io/tests-as-linear/)). But still, these simple models will be the occasion to introduce a more complex index, such as the **Bayes factor**.
+But before we do that, let us take a moment to remind ourselves and appreciate the fact that **all basic statistical pocedures** such as correlations, *t*-tests, ANOVAs or Chisquare tests ***are* linear regressions** (we strongly recommend  [this excellent demonstration](https://lindeloev.github.io/tests-as-linear/)). Nevertheless, these simple models will be the occasion to introduce a more complex index, such as the **Bayes factor**.
 
 ## Correlations
 
@@ -48,25 +48,25 @@ But before we do that, let us take a moment to remind ourselves and appreciate t
 
 Let us start, again, with a **frequentist correlation** between two continuous variables, the **width** and the **length** of the sepals of some flowers. The data is available in R as the `iris`  dataset (the same that was used in the [previous tutorial](https://easystats.github.io/bayestestR/articles/example1.html)). 
 
-Let's compute a Pearson's correlation test, store the results in an object called `result`, then display it:
+We will compute a Pearson's correlation test, store the results in an object called `result`, then display it:
 
 ```{r message=FALSE, warning=FALSE, eval=TRUE}
 result <- cor.test(iris$Sepal.Width, iris$Sepal.Length)
 result
 ```
 
-As you can see in the output, the test that we did actually compared two hypotheses: the **null hypothesis** (no correlation) with the **alternative hypothesis** (a non-null correlation). Based on the *p*-value, the null hypothesis cannot be rejected: the correlation between the two variables is **negative but not significant** (r = -.12, p > .05).
+As you can see in the output, the test that we did actually compared two hypotheses: the **null hypothesis** (*h0*; no correlation) with the **alternative hypothesis** (*h1*; a non-null correlation). Based on the *p*-value, the null hypothesis cannot be rejected: the correlation between the two variables is **negative but not significant** (r = -.12, p > .05).
 
 ### Bayesian correlation 
 
-To compute a Bayesian correlation test, we will need the [`BayesFactor`](https://richarddmorey.github.io/BayesFactor/) package (you can install it by running `install.packages("BayesFactor")`). We will then load this package, compute the correlation using the `correlationBF()` function and store the results in a similar fashion. 
+To compute a Bayesian correlation test, we will need the [`BayesFactor`](https://richarddmorey.github.io/BayesFactor/) package (you can install it by running `install.packages("BayesFactor")`). We can then load this package, compute the correlation using the `correlationBF()` function and store the results in a similar fashion. 
  
 ```{r message=FALSE, warning=FALSE, results='hide'}
[…]
