+++ vignettes/example3.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Example 3: Become a Bayesian master"
+title: "3. Become a Bayesian master"
 output: 
   github_document:
     toc: true
@@ -40,6 +40,13 @@ knitr::include_graphics("https://github.com/easystats/easystats/raw/master/man/f
 ```
 
 
+
+
+## Mixed Models
+
+TO BE CONTINUED.
+
+### Priors
 
 TO BE CONTINUED.
 
