### 2019-07-24 db4b1cb
new function: reshape_ci
