+++ R/utils_hdi_ci.R
@@ -178,3 +178,62 @@
   class(dat) <- unique(c(paste0("bayestestR_", fun), paste0("see_", fun), class(dat)))
   dat
 }
+
+
+
+
+
+
+#' Reshape CI between wide/long formats
+#'
+#' Reshape CI between wide/long formats.
+#'
+#' @param x A data.frame containing \code{CI_low} and \code{CI_high}.
+#'
+#' @examples
+#' x <- data.frame(replicate(4, rnorm(100)))
+#' x <- ci(x, ci = c(0.68, 0.89, 0.95))
+#' reshape_ci(x)
+#' reshape_ci(reshape_ci(x))
+#' @export
+reshape_ci <- function(x){
+  if("CI_low" %in% names(x) & "CI_high" %in% names(x)){
+    if(length(unique(x$CI)) > 1){
+      x <- reshape(x,
+                   idvar = "Parameter",
+                   timevar = "CI",
+                   direction="wide",
+                   sep = "_")
+      row.names(x) <- NULL
+    }
+  } else{
+    lows <- grepl("CI_low_*", names(x))
+    highs <- grepl("CI_high_*", names(x))
+    ci <- as.numeric(gsub("CI_low_", "", names(x)[lows]))
+    if(paste0(ci, collapse="-") != paste0(gsub("CI_high_", "", names(x)[highs]), collapse="-")){
+      stop("Something went wrong in the CIs reshaping.")
+      return(x)
+    }
+    if(sum(lows) > 1 & sum(highs) > 1){
+      low <- reshape(x[!highs],
+              direction="long",
+              varying=list(names(x)[lows]),
+              sep = "_",
+              timevar="CI",
+              v.names="CI_low",
+              times = ci)
+      high <- reshape(x[!lows],
+                     direction="long",
+                     varying=list(names(x)[highs]),
+                     sep = "_",
+                     timevar="CI",
+                     v.names="CI_high",
+                     times = ci)
+      x <- merge(low, high)
+      x$id <- NULL
+      x <- x[order(x$Parameter),]
+      row.names(x) <- NULL
+    }
+  }
+  x
+}
