+++ vignettes/indicesEstimationComparison.Rmd
@@ -80,7 +80,7 @@ df <- read.csv("https://raw.github.com/easystats/circus/master/data/bayesSim_stu
 
 ```{r, message=FALSE, warning=FALSE}
 df %>%
-  select(error, true_effect, outcome_type, beta, Median, Mean, MAP) %>%
+  select(error, true_effect, outcome_type, Coefficient, Median, Mean, MAP) %>%
   gather(estimate, value, -error, -true_effect, -outcome_type) %>%
   mutate(temp = as.factor(cut(error, 10, labels = FALSE))) %>% 
   group_by(temp) %>% 
@@ -93,7 +93,7 @@ df %>%
   # geom_smooth(method="loess") +
   geom_boxplot(outlier.shape=NA) +
   theme_modern() +
-  scale_fill_manual(values = c("beta" = "#607D8B", "MAP" = "#795548", "Mean" = "#FF9800", "Median" = "#FFEB3B"),
+  scale_fill_manual(values = c("Coefficient" = "#607D8B", "MAP" = "#795548", "Mean" = "#FF9800", "Median" = "#FFEB3B"),
                     name = "Index") +
   ylab("Point-estimate") +
   xlab("Noise") +
@@ -105,7 +105,7 @@ df %>%
 
 ```{r, message=FALSE, warning=FALSE}
 df %>%
-  select(sample_size, true_effect, outcome_type, beta, Median, Mean, MAP) %>%
+  select(sample_size, true_effect, outcome_type, Coefficient, Median, Mean, MAP) %>%
   gather(estimate, value, -sample_size, -true_effect, -outcome_type) %>%
   mutate(temp = as.factor(cut(sample_size, 10, labels = FALSE))) %>% 
   group_by(temp) %>% 
@@ -118,7 +118,7 @@ df %>%
   # geom_smooth(method="loess") +
   geom_boxplot(outlier.shape=NA) +
   theme_modern() +
-  scale_fill_manual(values = c("beta" = "#607D8B", "MAP" = "#795548", "Mean" = "#FF9800", "Median" = "#FFEB3B"),
+  scale_fill_manual(values = c("Coefficient" = "#607D8B", "MAP" = "#795548", "Mean" = "#FF9800", "Median" = "#FFEB3B"),
                     name = "Index") +
   ylab("Point-estimate") +
   xlab("Sample size") +
@@ -133,7 +133,7 @@ We fitted a (frequentist) multiple linear regression to statistically test the t
 
 ```{r, message=FALSE, warning=FALSE}
 df %>%
-  select(sample_size, error, true_effect, outcome_type, beta, Median, Mean, MAP) %>%
+  select(sample_size, error, true_effect, outcome_type, Coefficient, Median, Mean, MAP) %>%
   gather(estimate, value, -sample_size, -error, -true_effect, -outcome_type) %>%
   glm(true_effect ~ outcome_type / estimate / value, data=., family="binomial") %>%
   broom::tidy() %>%
@@ -147,10 +147,10 @@ df %>%
 
 This suggests that, in order to delineate between the presence and the absence of an effect, compared to the frequentist's beta:
 
-- For linear models, the **Mean** was the better predictor, closely followed by the **Median**, the **MAP** and the frequentist **beta**.
-- For logistic models, the **MAP** was the better predictor, followed by the **Median**, the **Mean** and the frequentist **beta**.
+- For linear models, the **Mean** was the better predictor, closely followed by the **Median**, the **MAP** and the frequentist **Coefficient**.
+- For logistic models, the **MAP** was the better predictor, followed by the **Median**, the **Mean** and, behind, the frequentist **Coefficient**.
 
[…]
