+++ vignettes/indicesExistenceComparison.Rmd
@@ -33,7 +33,8 @@ csl: apa.csl
 ```{r message=FALSE, warning=FALSE, include=FALSE}
 library(knitr)
 options(knitr.kable.NA = '')
-knitr::opts_chunk$set(comment=">")
+knitr::opts_chunk$set(comment=">",
+  dpi=100)
 options(digits=2)
 ```
 
@@ -95,6 +96,7 @@ library(ggplot2)
 library(dplyr)
 library(tidyr)
 library(see)
+library(ggExtra)
 
 df <- read.csv("https://raw.github.com/easystats/circus/master/data/bayesSim_study1.csv")
 ```
@@ -109,39 +111,44 @@ df <- read.csv("https://raw.github.com/easystats/circus/master/data/bayesSim_stu
 
 ##### Sensitivity to Noise
 
-```{r, message=FALSE, warning=FALSE, fig.height=25, fig.width=15}
-df %>%
-  select(outcome_type, true_effect, error, sample_size, p_value, p_direction, p_MAP, p_ROPE, ROPE_89, ROPE_95, ROPE_full, BF_log) %>%
+```{r, message=FALSE, warning=FALSE, fig.height=25, fig.width=15, eval=FALSE}
+figure1 <- df %>%
+  select(outcome_type, true_effect, error, sample_size, p_value, p_direction, p_MAP, ROPE_95, ROPE_full, BF_log) %>%
   gather(index, value, -error, -sample_size, -true_effect, -outcome_type) %>%
   mutate(true_effect = as.factor(true_effect),
-         index = factor(index, levels=c("p_value", "p_direction", "p_MAP", "p_ROPE", "ROPE_89", "ROPE_95", "ROPE_full", "BF_log"))) %>%
-  mutate(temp = as.factor(cut(error, 10, labels = FALSE))) %>% 
+         index = factor(index, levels=c("p_value", "p_direction", "p_MAP", "ROPE_95", "ROPE_full", "BF_log"))) %>%
+  mutate(temp = as.factor(cut(error, 8, labels = FALSE))) %>% 
   group_by(temp) %>% 
   mutate(error_group = round(mean(error), 1)) %>% 
   ungroup() %>% 
   ggplot(aes(x = error_group, y = value, fill = index, colour=true_effect, group = interaction(index, true_effect, error_group))) +
-  # geom_jitter(shape=16, alpha=0.02) +
   geom_boxplot(outlier.shape = NA) +
   facet_wrap(~index * outcome_type, scales = "free", ncol=2) +
   theme_modern() +
-  scale_color_manual(values = c(`0` = "#f44336", `1` = "#8BC34A", name="Effect")) +
+  scale_color_manual(values = c(`0` = "#f44336", `1` = "#8BC34A"), name="Effect") +
   scale_fill_manual(values = c("p_value"="#607D8B", "p_MAP" = "#4CAF50", "p_direction" = "#2196F3",
-                               "ROPE_89" = "#FFC107", "ROPE_95" = "#FF9800", "ROPE_full" = "#FF5722",
-                               "p_ROPE"="#E91E63", "BF_log"="#9C27B0"), guide=FALSE) +
+                               "ROPE_95" = "#FFC107", "ROPE_full" = "#FF9800", "BF_log"="#E91E63"), 
+                    guide=FALSE) +
   ylab("Index Value") +
   xlab("Noise")
+figure1
 ```
+```{r, message=FALSE, warning=FALSE, fig.cap="Figure 1.", fig.align='center', out.width="100%", echo=FALSE}
+ggsave("2019_frontiers/figure1.png", figure1, width=21*0.5, height=29.7*0.5, dpi=600)
+ggsave("2019_frontiers/figure1_small.png", figure1, width=21*0.5, height=29.7*0.5, dpi=150)
 
[…]
