+++ vignettes/indicesExistenceComparison.Rmd
@@ -77,9 +77,7 @@ For all of the simulated models, we computed the following indices:
 The Region of Practical Equivalence (ROPE) was defined as ranging from -0.1 to 0.1.
 
 
-## Study 1: Relationship with Noise and Sample Size
-
-### Methods
+## Methods
 
 
 The simulation aimed at modulating the following characteristics:
@@ -104,14 +102,14 @@ df <- read.csv("https://raw.github.com/easystats/circus/master/data/bayesSim_stu
 
 
 
-### Results
+## Results
 
-#### Effect Detection
+### Effect Detection
 
 
-##### Sensitivity to Noise
+#### Sensitivity to Noise
 
-```{r, message=FALSE, warning=FALSE, fig.height=25, fig.width=15, eval=FALSE}
+```{r, message=FALSE, warning=FALSE, eval=FALSE}
 figure1 <- df %>%
   select(outcome_type, true_effect, error, sample_size, p_value, p_direction, p_MAP, ROPE_95, ROPE_full, BF_log) %>%
   gather(index, value, -error, -sample_size, -true_effect, -outcome_type) %>%
@@ -134,15 +132,15 @@ figure1 <- df %>%
 figure1
 ```
 ```{r, message=FALSE, warning=FALSE, fig.cap="Figure 1.", fig.align='center', out.width="100%", echo=FALSE}
-ggsave("2019_frontiers/figure1.png", figure1, width=21*0.5, height=29.7*0.5, dpi=600)
-ggsave("2019_frontiers/figure1_small.png", figure1, width=21*0.5, height=29.7*0.5, dpi=150)
+# ggsave("2019_frontiers/figure1.png", figure1, width=21*0.5, height=29.7*0.5, dpi=600)
+# ggsave("2019_frontiers/figure1_small.png", figure1, width=21*0.5, height=29.7*0.5, dpi=100)
 
 knitr::include_graphics("https://github.com/easystats/bayestestR/raw/master/vignettes/2019_frontiers/figure1_small.png")
 ```
 
-##### Sensitivity to Sample Size
+#### Sensitivity to Sample Size
 
-```{r, message=FALSE, warning=FALSE, fig.height=25, fig.width=15, eval=FALSE}
+```{r, message=FALSE, warning=FALSE, eval=FALSE}
 figure2 <- df %>%
   select(outcome_type, true_effect, error, sample_size, p_value, p_direction, p_MAP, ROPE_95, ROPE_full, BF_log) %>%
   gather(index, value, -error, -sample_size, -true_effect, -outcome_type) %>%
@@ -165,16 +163,16 @@ figure2 <- df %>%
   xlab("Sample Size")
 figure2
 ```
-```{r, message=FALSE, warning=FALSE, fig.cap="Figure 1.", fig.align='center', out.width="100%", echo=FALSE}
-ggsave("2019_frontiers/figure2.png", figure2, width=21*0.5, height=29.7*0.5, dpi=600)
-ggsave("2019_frontiers/figure2_small.png", figure2, width=21*0.5, height=29.7*0.5, dpi=150)
+```{r, message=FALSE, warning=FALSE, fig.cap="Figure 2.", fig.align='center', out.width="100%", echo=FALSE}
+# ggsave("2019_frontiers/figure2.png", figure2, width=21*0.5, height=29.7*0.5, dpi=600)
+# ggsave("2019_frontiers/figure2_small.png", figure2, width=21*0.5, height=29.7*0.5, dpi=100)
 
 knitr::include_graphics("https://github.com/easystats/bayestestR/raw/master/vignettes/2019_frontiers/figure2_small.png")
 ```
 
 
 
-##### Statistical Modelling
+#### Statistical Modelling
 
[…]
