+++ R/p_rope.R
@@ -23,7 +23,7 @@
 #' p_rope(df, precision = 1)
 #'
 #' library(rstanarm)
-#' model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200)
+#' model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200, refresh = 0)
 #' p_rope(model, precision = 1)
 #'
 #' library(emmeans)
