### 2019-07-24 8d85237
had to move bayesian indices vignette to other repo to fix size issue
