+++ vignettes/probability_of_direction.Rmd
@@ -51,39 +51,40 @@ Indices of effect existence, such as the *pd*, are particularly useful in explor
 
 # Relationship with the *p*-value
 
-In most cases, it seems that the *pd* has a direct correspondance with the frequentist **one-sided *p*-value** through the formula: $$p_{one sided} = 1-p_d/100$$ Similarly, the **two-sided *p*-value** (the most commonly reported one) is equivalent through the formula: $$p_{two sided} = 2*(1-p_d/100)$$ Thus, the two-sided *p*-value of respectively **.1**, **.05**, **.01** and **.001** would correspond approximately to a *pd* of **95\%**, **97.5\%**, **99.5\%** and **99.95\%** .
+In most cases, it seems that the *pd* has a direct correspondence with the frequentist **one-sided *p*-value** through the formula: $$p_{one sided} = 1-p_d/100$$ Similarly, the **two-sided *p*-value** (the most commonly reported one) is equivalent through the formula: $$p_{two sided} = 2*(1-p_d/100)$$ Thus, the two-sided *p*-value of respectively **.1**, **.05**, **.01** and **.001** would correspond approximately to a *pd* of **95\%**, **97.5\%**, **99.5\%** and **99.95\%** .
 
 
 ```{r message=FALSE, warning=FALSE, echo=FALSE, fig.cap="Correlation between the frequentist p-value and the probability of direction (pd)", fig.align='center'}
 library(dplyr)
 library(tidyr)
 library(ggplot2)
+library(see)
 
 read.csv("https://raw.github.com/easystats/circus/master/data/bayesSim_study1.csv") %>% 
   mutate(effect_existence = ifelse(true_effect == 1, "Presence of true effect", "Absence of true effect")) %>% 
   ggplot(aes(x=p_direction, y=p_value, color=effect_existence)) +
-  geom_point(alpha=0.1, stroke=0, shape=16) +
-  geom_segment(aes(x=95, y=Inf, xend=95, yend=0.1), color="black", linetype="longdash") +
-  geom_segment(aes(x=-Inf, y=0.1, xend=95, yend=0.1), color="black", linetype="longdash") +
-  theme_classic() +
+  geom_point2(alpha=0.1) +
+  geom_segment(aes(x=0.95, y=Inf, xend=0.95, yend=0.1), color="black", linetype="longdash") +
+  geom_segment(aes(x=-Inf, y=0.1, xend=0.95, yend=0.1), color="black", linetype="longdash") +
+  theme_modern() +
   scale_y_reverse(breaks = round(seq(0, 1, length.out = 11), digits=2)) +
-  scale_x_continuous(breaks = round(c(95, seq(50, 100, length.out = 6)))) +
+  scale_x_continuous(breaks = round(c(0.95, seq(0.5, 1, length.out = 6)), 2)) +
   scale_color_manual(values=c("Presence of true effect"="green", "Absence of true effect"="red")) +
   theme(legend.title = element_blank()) +
   guides(colour = guide_legend(override.aes = list(alpha = 1))) +
-  xlab("\nProbability of Direction (pd)") +
-  ylab("Frequentist p-value\n")
+  xlab("Probability of Direction (pd)") +
+  ylab("Frequentist p-value")
 ```
 
 
 
 > **But if it's like the *p*-value, it must be bad because the *p*-value is bad [*insert reference to the reproducibility crisis*].**
 
[…]
