+++ man/eti.Rd
@@ -103,7 +103,7 @@ eti(df)
 eti(df, ci = c(.80, .89, .95))
 
 library(rstanarm)
-model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200)
+model <- stan_glm(mpg ~ wt + gear, data = mtcars, chains = 2, iter = 200, refresh = 0)
 eti(model)
 eti(model, ci = c(.80, .89, .95))
 
