### 2019-07-25 db913bc
add section in vignette by pd = 100% edgecase
