+++ vignettes/bibliography.bib
@@ -261,6 +261,17 @@
   publisher={American Psychological Association}
 }
 
+@article{lilienfeld2015fifty,
+  doi = {10.3389/fpsyg.2015.01100},
+  title={Fifty psychological and psychiatric terms to avoid: a list of inaccurate, misleading, misused, ambiguous, and logically confused words and phrases},
+  author={Lilienfeld, Scott O and Sauvign{\'e}, Katheryn C and Lynn, Steven Jay and Cautin, Robin L and Latzman, Robert D and Waldman, Irwin D},
+  journal={Frontiers in Psychology},
+  volume={6},
+  pages={1100},
+  year={2015},
+  publisher={Frontiers}
+}
+
 
 @misc{mcelreath2014rethinking,
   title={rethinking: Statistical Rethinking book package. R package version 1.391},
