### 2019-07-25 de652af
make reshape_ci smarter
