### 2019-07-25 b91048e
minor docs
