+++ man/reshape_ci.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils_hdi_ci.R
+% Please edit documentation in R/reshape_ci.R
 \name{reshape_ci}
 \alias{reshape_ci}
 \title{Reshape CI between wide/long formats}
@@ -13,8 +13,15 @@ reshape_ci(x)
 Reshape CI between wide/long formats.
 }
 \examples{
+library(bayestestR)
+
 x <- data.frame(replicate(4, rnorm(100)))
 x <- ci(x, ci = c(0.68, 0.89, 0.95))
+reshape_ci(x)
+reshape_ci(reshape_ci(x))
+
+x <- data.frame(replicate(4, rnorm(100)))
+x <- describe_posterior(x, ci = c(0.68, 0.89, 0.95))
 reshape_ci(x)
 reshape_ci(reshape_ci(x))
 }
