### 2019-07-25 #1541 @BarryDeCicco
Please take me off of the list of reviewers. I have received over a dozen e-mails on four projects today.
The first project, yesterday, generated ~30 e-mails to me, most of them not understandable.IMHO, your process for reviewing needs serious work.
Sincerely,
Barry
    On Thursday, July 25, 2019, 6:38:25 AM EDT, whedon <notifications@github.com> wrote:

 👉 Check article proof 📄 👈
