### 2019-07-25 e4b14cc
better doc for priors in BF
