+++ R/bayesfactor_parameters.R
@@ -23,7 +23,7 @@
 #' (for a interval-null).
 #' @inheritParams hdi
 #'
-#' @return A data frame containing the Bayes factor representing evidence \emph{against} the (point) null effect model.
+#' @return A data frame containing the Bayes factor representing evidence \emph{against} the null.
 #'
 #' @details This method is used to compute Bayes factors based on prior and posterior distributions.
 #' When \code{posterior} is a model (\code{stanreg}, \code{brmsfit}), posterior and prior samples are
@@ -36,9 +36,15 @@
 #' It is important to provide the correct \code{prior} for meaningful results.
 #' \itemize{
 #'   \item When \code{posterior} is a numerical vector, \code{prior} should also be a numerical vector.
-#'   \item When \code{posterior} is an \code{emmGrid} object based on a \code{stanreg} or \code{brmsfit} model, \code{prior} should be \emph{that model object} (see example).
-#'   \item When \code{posterior} is a \code{stanreg} or \code{brmsfit} model, there is no need to specify \code{prior}, as prior samples are drawn internally.
 #'   \item When \code{posterior} is a \code{data.frame}, \code{prior} should also be a \code{data.frame}, with matching column order.
+#'   \item When \code{posterior} is a \code{stanreg} or \code{brmsfit} model: \itemize{
+#'     \item \code{prior} can be set to \code{NULL}, in which case prior samples are drawn internally.
+#'     \item \code{prior} can also be a model equvilant to \code{posterior} but with samples from the priors \emph{only}.
+#'   }
+#'   \item When \code{posterior} is an \code{emmGrid} object: \itemize{
+#'     \item \code{prior} should be the \code{stanreg} or \code{brmsfit} model used to create the \code{emmGrid} objects.
+#'     \item \code{prior} can also be an \code{emmGrid} object equvilant to \code{posterior} but created with a model of priors samples \emph{only}.
+#'   }
 #' }}
 #' \subsection{One-sided Tests (setting an order restriction)}{
 #' One sided tests (controlled by \code{direction}) are conducted by restricting the prior and
@@ -201,7 +207,7 @@ bayesfactor_parameters.emmGrid <- function(posterior, prior = NULL,
       "Prior not specified! ",
       "Please provide the original model to get meaningful results."
     )
-  } else {
+  } else if (!inherits(prior, "emmGrid")) { # then is it a model
     prior <- .update_to_priors(prior, verbose = verbose)
     prior <- insight::get_parameters(prior, effects = "fixed")
     prior <- stats::update(posterior, post.beta = as.matrix(prior))
