### 2019-07-25 ec66420
hopefully final fixes
