+++ R/utils.R
@@ -77,3 +77,26 @@
 .select_nums <- function(x) {
   x[unlist(lapply(x, is.numeric))]
 }
+
+
+
+#' Used in describe_posterior
+#' @keywords internal
+.reoder_rows <- function(x, out, ci = NULL) {
+
+  if(!is.data.frame(out) || nrow(out) == 1){
+    return(out)
+  }
+
+  if(is.null(ci)){
+    refdata <- point_estimate(x, centrality = "median", dispersion = FALSE)
+    order <- refdata$Parameter
+    out <- out[match(order, out$Parameter), ]
+  } else{
+    uncertainty <- ci(x, ci = ci)
+    order <- paste0(uncertainty$Parameter, uncertainty$CI)
+    out <- out[match(order, paste0(out$Parameter, out$CI)), ]
+  }
+  rownames(out) <- NULL
+  out
+}
