### 2019-07-25 74bb716
deactivate link to BayesFactor::priorOdds<-
Otherwise it igives this warning when building the pkg:

`Rd warning: easystats/bayestestR/man/bayesfactor_inclusion.Rd:15: file link 'priorOdds<-' in package 'BayesFactor' does not exist and so has been treated as a topic`

I guess it is because of the weird form of this call (with the assign operator).
