### 2019-07-25 347beb6
news-style, add () to funs
