### 2019-07-25 30c8a00
refine deprecated function message
