### 2019-07-26 0ec2fe9
fix BF tests and docs
