+++ man/bayesfactor.Rd
@@ -20,7 +20,7 @@ bayesfactor(..., prior = NULL, direction = "two-sided", null = 0,
 \item{null}{Value of the null, either a scaler (for point-null) or a a range
 (for a interval-null).}
 
-\item{hypothesis}{A character vector specifying the restrictions as logical conditions (see examples below).}
+\item{hypothesis}{Deprecated in favour of \code{null}.}
 
 \item{effects}{Should results for fixed effects, random effects or both be returned?
 Only applies to mixed models. May be abbreviated.}
