### 2019-07-26 badeb10
Update test-bayesfactor_parameters.R
