### 2019-07-26 8ee1b50
hypothesis before prior in BF_restriced
