### 2019-07-26 a26ee76
change class for plot
