+++ NAMESPACE
@@ -97,9 +97,9 @@ S3method(p_rope,numeric)
 S3method(p_rope,stanreg)
 S3method(plot,bayesfactor_models)
 S3method(plot,bayesfactor_parameters)
-S3method(plot,bayestestR_ci)
 S3method(plot,equivalence_test)
 S3method(plot,estimate_density)
+S3method(plot,eti)
 S3method(plot,hdi)
 S3method(plot,p_direction)
 S3method(plot,point_estimate)
