### 2019-07-28 #1541 @cMadan
@BarryDeCicco, you should unsubscribe from the GitHub notifications for this repository, or otherwise change your notification settings. See the second post (https://github.com/openjournals/joss-reviews/issues/1541#issuecomment-507707086) in this review thread (same information is available in every review thread).
