### 2019-07-29 #1541 @tjmahr
I worked with version 0.2.4, the most recent on the GitHub repository, although the version mentioned in the checklist is 0.2.3

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I do not see anything for contribution guidelines. I would add a `CONTRIBUTING.md` file.

## Comments

The amount of documentation to support the package is very generous. For my review, however, I focused on the README and the software paper.

This package borrows a lot from the Kruschke school of Bayesian inference. HDI and ROPE are a distinct feature of his tutorials and textbook; one does not seem them very often in works by, say, Stan developers. Therefore, this package is tremendously useful for those reading his tutorials, or for people like me, who occasionally will quantify an effect with a ROPE percentage or who want to learn about using Bayes factors.

The ROPE procedure and other indices use the highest density interval. Is there any option to use an equal-tailed interval?

### README

I was confused by the README. When I see R code followed immediately by a plot, I assume that the R code produced the plot. But the functions in the README produce text output (which is not included in the README) and they do not produce plots. I would include the text output of the R code. I would also note that the figures there are diagrams meant to illustrate the statistical concept. The software paper does a good job of making this point clear.

I don't see a demo for `eti()`.

> Moreover, 89 is the highest prime number that does not exceed the already unstable 95% threshold (McElreath, 2015).

The primeness of 89 is not important. McElreath's choice of 89 in *Statistical Rethinking* text was to illustrate that interval widths are arbitrary and that there is nothing special about 95 or 90 compared to 89.

> equivalence_test() a Test for Practical Equivalence based on the

Needs a verb.

I don't understand the Bayes Factor diagram in the README.

> a range of -0.05 to -0.05.

This range is the same number twice.

> Savage-Dickey density ratio is computed

Should this have a reference?

> Probability of a Value

`density_at()` isn't doing computing a probability. I would remove `estimate_probability()` and `probability_at()` because they are just aliases for density functions and *density* is the more appropriate term.

I don't see a demo for the area under the curve functions.

### Documentation

> A ROPE-based p of 97% means that there is a probability of .97 that a parameter (described by its posterior distribution) is outside the ROPE. On the contrary, a ROPE-based p of -97% means that there is a probability of .97 that the parameter is inside the ROPE. (`R/p_rope.R`)

I don't understand how a p-value can get a negative percentage. What would a 0% p-value mean? If this index doesn't act like a familiar p-value, it is probably the wrong name for it.

### Software paper

The first mention of bayestestR in the second paragraph is awkward. Specifically, the text shifts from talking about common ways to describe effects in a Bayesian framework to talking about the features of the package:

> Additionally, bayestestR also focuses on implementing a Bayesian null-hypothesis testing framework ...

It's great that the output of `point_estimate()` prints out mean/median/map to make it clear what value is being used.

> However, bayestestR functions also include plotting capabilities via the see package (Lüdecke, Waggoner, Ben-Shachar, & Makowski, 2019).

I don't see any plotting examples in the README or documentation pages. I see plotting methods in the NAMESPACE.

I think it would worthwhile to demonstrate that the functions demoed in the article also work on models. For example, I can call `p_direction()` and `bayesfactor_parameters()` directly on a model and get the results for each parameter. One of the key contributions of this package is that it can make these indices immediately available to users who are comfortable with particular brms and rstanarm.

#### Proofreading concerns

Every reference of Kruschke spells out the author's full name.

Figure 2 should be referenced in the text.

> (i.e. the difference

Needs a comma.

> The Bayesian framework allows to neatly delineate

Allows *one*.

> developped

> Nevertheless, in the absence of user-provided values, bayestestR will automatically find an appropriate range

*Nevertheless* doesn't make sense.

> bases on prior and posterior samples

Based.

The system for building the references section should protect some words from being converted to lowercase. (In LaTeX, this is done with {}). Right now, for example, it says `Brms: An r package for bayesian multilevel models using stan` but I would make sure that the system produces `brms: An R package for Bayesian multilevel models using Stan`.
