### 2019-07-30 98fa54c
protect brms reference
