+++ paper/paper.bib
@@ -218,7 +218,7 @@
 	volume = {80},
 	number = {1},
 	author = {Paul-Christian Bürkner},
-	title = {brms: An R Package for Bayesian Multilevel Models Using Stan},
+	title = {{brms}: An R Package for Bayesian Multilevel Models Using Stan},
 	journal = {Journal of Statistical Software}
 }
 
