### 2019-07-30 feae6af
hopefully fix Kruschke's fullname (does anyone has any idea what would that happen tho?)
