+++ paper/paper.bib
@@ -52,7 +52,7 @@
 	volume = {1},
 	number = {2},
 	pages = {270--280},
-	author = {John K. Kruschke},
+	author = {Kruschke, John K},
 	title = {Rejecting or Accepting Parameter Values in Bayesian Estimation},
 	journal = {Advances in Methods and Practices in Psychological Science}
 }
@@ -363,7 +363,7 @@
 	year = 2015,
 	publisher = {Elsevier},
 	pages = {399--416},
-	author = {John K. Kruschke},
+	author = {Kruschke, John K},
 	title = {Stan},
 	booktitle = {Doing Bayesian Data Analysis}
 }
@@ -487,7 +487,7 @@
 	volume = {15},
 	number = {4},
 	pages = {722--752},
-	author = {John K. Kruschke and Herman Aguinis and Harry Joo},
+	author = {Kruschke, John K and Herman Aguinis and Harry Joo},
 	title = {The Time Has Come: Bayesian methods for data analysis in the organizational sciences},
 	journal = {Organizational Research Methods}
 }
@@ -543,7 +543,7 @@
 	volume = {14},
 	number = {7},
 	pages = {293--300},
-	author = {John K. Kruschke},
+	author = {Kruschke, John K},
 	title = {What to believe: Bayesian methods for data analysis},
 	journal = {Trends in Cognitive Sciences}
 }
