+++ paper/paper.md
@@ -49,7 +49,7 @@ The following demonstration of functions is accompanied by figures to illustrate
 
 **bayestestR** offers two functions to compute point-estimates from posterior distributions: `map_estimate()` and `point_estimate()`, the latter providing options to calculate the mean, median or MAP estimate of a posterior distribution. [`map_estimate()`](https://easystats.github.io/bayestestR/reference/map_estimate.html) is a convenient function to calculate the **Maximum A Posteriori** (MAP) estimate directly.
 
-The **posterior mean** minimizes expected _squared_ error, whereas the **posterior median** minimizes expected _absolute_ error (i.e. the difference of estimates from true values over samples). The **MAP estimate** corresponds to the most probable value of a posterior distribution.
+The **posterior mean** minimizes expected _squared_ error, whereas the **posterior median** minimizes expected _absolute_ error (i.e., the difference of estimates from true values over samples). The **MAP estimate** corresponds to the most probable value of a posterior distribution.
 
 ``` r
 set.seed(1)
@@ -96,14 +96,14 @@ eti(posterior)
 
 ## Null-Hypothesis Significance Testing (NHST)
 
-The Bayesian framework allows to neatly delineate and quantify different aspects of hypothesis testing, including effect *existence* and *significance*, and different indices have been developped to describe them.
+The Bayesian framework allows one to neatly delineate and quantify different aspects of hypothesis testing, including effect *existence* and *significance*, and different indices have been developed to describe them.
 
 
 ### ROPE and Test for Practical Equivalence
 
 [`rope()`](https://easystats.github.io/bayestestR/reference/rope.html) computes the proportion of the HDI (default to the 89% HDI) of a posterior distribution that lies within a Region Of Practical Equivalence (the **ROPE**; see figure 3, panel A).
 
-Statistically, the probability of a posterior distribution of being different from 0 does not make much sense (the probability of it being different from a single point being infinite). Therefore, the idea underlining ROPE is to let the user define an area around the null value enclosing values that are *equivalent to the null* value for practical purposes [@kruschke2018bayesian; @kruschke2018rejecting]. Nevertheless, in the absence of user-provided values, **bayestestR** will automatically find an appropriate range for the ROPE using the  [`rope_range()`](https://easystats.github.io/bayestestR/reference/rope_range.html) function.
+Statistically, the probability of a posterior distribution of being different from 0 does not make much sense (the probability of it being different from a single point being infinite). Therefore, the idea underlining ROPE is to let the user define an area around the null value enclosing values that are *equivalent to the null* value for practical purposes [@kruschke2018bayesian; @kruschke2018rejecting]. In the absence of user-provided values, **bayestestR** will automatically find an appropriate range for the ROPE using the  [`rope_range()`](https://easystats.github.io/bayestestR/reference/rope_range.html) function.
 
 ``` r
 rope(distribution_normal(1000, mean = 1), range = c(-0.5, 0.5))
@@ -147,7 +147,7 @@ p_direction(distribution_normal(100, 0.4, 0.2))
 
 ### Bayes Factor
 
-[**`bayesfactor_parameters()`**](https://easystats.github.io/bayestestR/reference/bayesfactor_parameters.html) computes Bayes factors against the null (either a point or an interval), bases on prior and posterior samples of a single parameter. This Bayes factor indicates the degree by which the mass of the posterior distribution has shifted further away from or closer to the null value(s) (relative to the prior distribution), thus indicating if the null value has become less or more likely given the observed data. 
[…]
