+++ paper/paper.md
@@ -47,7 +47,7 @@ The following demonstration of functions is accompanied by figures to illustrate
 
 ## Indices of Centrality: Point-estimates
 
-**bayestestR** offers two functions to compute point-estimates from posterior distributions: `map_estimate()` and `point_estimate()`, the latter providing options to calculate the mean, median or MAP estimate of a posterior distribution. [`map_estimate()`](https://easystats.github.io/bayestestR/reference/map_estimate.html) is a convenient function to calculate the **Maximum A Posteriori** (MAP) estimate directly.
+**bayestestR** offers two functions to compute point-estimates from posterior distributions: `map_estimate()` and `point_estimate()`, the latter providing options to calculate the mean, median or MAP estimate of a posterior distribution (see **Figure 1**). [`map_estimate()`](https://easystats.github.io/bayestestR/reference/map_estimate.html) is a convenient function to calculate the **Maximum A Posteriori** (MAP) estimate directly.
 
 The **posterior mean** minimizes expected _squared_ error, whereas the **posterior median** minimizes expected _absolute_ error (i.e., the difference of estimates from true values over samples). The **MAP estimate** corresponds to the most probable value of a posterior distribution.
 
@@ -71,10 +71,10 @@ point_estimate(posterior, centrality = "map")
 
 ## Quantifying Uncertainty: The Credible Interval (CI)
 
-In order to measure the uncertainty associated with the estimation, **bayestestR** provides two functions: `eti()`, the Equal-Tailed Interval **ETI**, and `hdi()`, the **Highest Density Interval (HDI)**. Both indices (accessible via the `method` argument in the `ci()` function) can be used in the context of Bayesian posterior characterisation as **Credible Interval (CI)**.
+In order to measure the uncertainty associated with the estimation, **bayestestR** provides two functions: `eti()`, the **Equal-Tailed Interval (ETI)**, and `hdi()`, the **Highest Density Interval (HDI)**. Both indices (accessible via the `method` argument in the `ci()` function) can be used in the context of Bayesian posterior characterisation as **Credible Interval (CI)**.
 
 [`hdi()`](https://easystats.github.io/bayestestR/reference/hdi.html)
-computes the HDI of a posterior distribution, i.e., the interval which contains all points within the interval have a higher probability density than points outside the interval. HDIs have a particular property: Unlike an equal-tailed interval (computed by [`eti()`](https://easystats.github.io/bayestestR/reference/eti.html)) that typically exclude 2.5% from each tail of the distribution, the HDI is *not* equal-tailed and therefore always includes the mode(s) of posterior distributions.
+computes the HDI of a posterior distribution, i.e., the interval which contains all points within the interval have a higher probability density than points outside the interval (see **Figure 2**). HDIs have a particular property: Unlike an equal-tailed interval (computed by [`eti()`](https://easystats.github.io/bayestestR/reference/eti.html)) that typically exclude 2.5% from each tail of the distribution, the HDI is *not* equal-tailed and therefore always includes the mode(s) of posterior distributions.
 
 By default, `hdi()` and `eti()` return the 89% intervals (`ci = 0.89`), deemed to be more stable than, for instance, 95% intervals. An effective sample size of at least 10.000 is recommended if 95% intervals should be computed [@kruschke2015doing]. Moreover, 89 is the highest prime number that does not exceed the already unstable and arbritrary 95% threshold [@mcelreath2018statistical].
 
@@ -101,7 +101,7 @@ The Bayesian framework allows one to neatly delineate and quantify different asp
 
 ### ROPE and Test for Practical Equivalence
 
[…]
