### 2019-07-30 169edc9
added bayestestR mention in first paragraph
