### 2019-07-30 c2ee166
added visualisation example and that functions can be applied to models
