+++ paper/paper.md
@@ -39,7 +39,7 @@ The Bayesian framework for statistics is quickly gaining in popularity among sci
 
 Effects in the Bayesian framework can be described by [characterizing their posterior distribution](https://easystats.github.io/bayestestR/articles/guidelines.html). Commonly reported indices include measures of centrality (e.g., the median, mean or MAP estimate) and uncertainty (the [*credible* interval](https://easystats.github.io/bayestestR/articles/credible_interval.html) - CI). Cum grano salis, these are considered the counterparts to the coefficient point-estimates and confidence intervals of the frequentist framework. Additionally, **bayestestR** also focuses on implementing a Bayesian null-hypothesis testing framework (in a Bayesian sense, i.e., extended to general testing of "effect existence") by providing access to both established and exploratory indices of effect *existence* and *significance* [such as the Bayes factor, @morey2011bayes; the ROPE, @kruschke2018bayesian; the MAP-based *p*-value, @mills2018objective, or the Probability of Direction - *pd*].
 
-Existing R packages allow users to easily fit a large variety of models and extract and visualize the posterior draws. However, most of these packages only return a limited set of indices (e.g., point-estimates and CIs). **bayestestR** provides a comprehensive and consistent set of functions to analyze and describe posterior distributions generated by a variety of models objects, including popular modeling packages such as **rstanarm** [@goodrich2018rstanarm], **brms** [@burkner2017brms] or **BayesFactor** [@morey2018bayesfactor], thus appearing as a useful tool supporting the usage and developpment of Bayesian statistics. The main functions are described below, and a full documentation is available on the [package's website](https://easystats.github.io/bayestestR).
+Existing R packages allow users to easily fit a large variety of models and extract and visualize the posterior draws. However, most of these packages only return a limited set of indices (e.g., point-estimates and CIs). **bayestestR** provides a comprehensive and consistent set of functions to analyze and describe posterior distributions generated by a variety of models objects, including popular modeling packages such as **rstanarm** [@goodrich2018rstanarm], **brms** [@burkner2017brms], **BayesFactor** [@morey2018bayesfactor] or **emmeans** [@lenth2019emmeans], thus appearing as a useful tool supporting the usage and developpment of Bayesian statistics. The main functions are described below, and a full documentation is available on the [package's website](https://easystats.github.io/bayestestR).
 
 # Examples of Features
 
@@ -55,15 +55,19 @@ The **posterior mean** minimizes expected _squared_ error, whereas the **posteri
 set.seed(1)
 posterior <- rchisq(100, 3)
 map_estimate(posterior)
+
 #> MAP = 1.46
 
 point_estimate(posterior)
+
 #> Median = 2.31
 
 point_estimate(posterior, centrality = "mean")
+
 #> Mean = 2.96
 
 point_estimate(posterior, centrality = "map")
+
 #> MAP = 1.46
 ```
 
@@ -80,12 +84,14 @@ By default, `hdi()` and `eti()` return the 89% intervals (`ci = 0.89`), deemed t
 
 ``` r
 hdi(posterior)
+
 #> # Highest Density Interval
 #> 
 #>       89% HDI
 #>  [0.11, 6.05]
 
 eti(posterior)
+
 #> # Equal-Tailed Interval
 #> 
 #>       89% ETI
@@ -107,6 +113,7 @@ Statistically, the probability of a posterior distribution of being different fr
 
[…]
