+++ paper/paper.bib
@@ -17,6 +17,14 @@
   pages = {1412},
 }
 
+@Manual{lenth2019emmeans,
+    title = {emmeans: Estimated Marginal Means, aka Least-Squares Means},
+    author = {Russell Lenth},
+    year = {2019},
+    note = {R package version 1.3.5.1},
+    url = {https://CRAN.R-project.org/package=emmeans},
+  }
+
 
 @book{mcelreath2018statistical,
 	doi = {10.1201/9781315372495},
