### 2019-07-30 a52de86
removed probability aliases and clarified docs
