+++ NAMESPACE
@@ -165,7 +165,6 @@ export(distribution_uniform)
 export(effective_sample)
 export(equivalence_test)
 export(estimate_density)
-export(estimate_probability)
 export(eti)
 export(hdi)
 export(map_estimate)
@@ -177,7 +176,6 @@ export(p_to_pd)
 export(pd)
 export(pd_to_p)
 export(point_estimate)
-export(probability_at)
 export(reshape_ci)
 export(rnorm_perfect)
 export(rope)
