### 2019-07-30 158e45a
explain BF fig
