+++ README.Rmd
@@ -259,6 +259,7 @@ posterior %>%
   xlab("\nParameter Value") +
   ylab("Probability Density\n")
 ```
+<center>*The lollipops represent the density of a point-null on the prior distribution (the blue lollipop on the dotted distribution) and on the posterior distribution (the red lollipop on the yellow distribution). The ratio between the two - the Svage-Dickey ratio - indicates the degree by which the mass of the parameter distribution has shifted away from or closer to the null.*</center>
 
 For more info, see [the Bayes factors vignette](https://easystats.github.io/bayestestR/articles/bayes_factors.html).
 
