### 2019-07-30 57c9eaf
add textual output to readme
