+++ README.Rmd
@@ -81,7 +81,7 @@ Click on the buttons above to access the package [**documentation**](https://eas
 
 [**`describe_posterior()`**](https://easystats.github.io/bayestestR/reference/describe_posterior.html) is the master function with which you can compute all of the indices cited below *at once*.
 
-```{r message=FALSE, warning=FALSE, results='hide'}
+```{r message=FALSE, warning=FALSE}
 describe_posterior(rnorm(1000))
 ```
 
@@ -93,8 +93,9 @@ describe_posterior(rnorm(1000))
 
 [**`map_estimate()`**](https://easystats.github.io/bayestestR/reference/map_estimate.html) find the **Highest Maximum A Posteriori (MAP)** estimate of a posterior, *i.e.*, the value associated with the highest probability density (the "peak" of the posterior distribution). In other words, it is an estimation of the *mode* for continuous parameters.
 
-```{r message=FALSE, warning=FALSE, results='hide'}
-map_estimate(rnorm(1000, 1, 1))
+```{r message=FALSE, warning=FALSE}
+posterior <- distribution_normal(100, 0.4, 0.2)
+map_estimate(posterior)
 ```
 
 ```{r message=FALSE, warning=FALSE, echo=FALSE}
@@ -126,8 +127,9 @@ Unlike equal-tailed intervals (see [`ci()`](https://easystats.github.io/bayestes
 
 By default, `hdi()` returns the 89\% intervals (`ci = 0.89`), deemed to be more stable than, for instance, 95\% intervals (Kruschke, 2014). An effective sample size of at least 10.000 is recommended if 95\% intervals should be computed (Kruschke 2014, p. 183ff). Moreover, 89 is the highest prime number that does not exceed the already unstable 95\% threshold (McElreath, 2015).
 
-```{r message=FALSE, warning=FALSE, results='hide'}
-hdi(rnorm(1000), ci = .89)
+```{r message=FALSE, warning=FALSE}
+posterior <- distribution_normal(100, 0.4, 0.2)
+hdi(posterior, ci = .89)
 ```
 
 ```{r message=FALSE, warning=FALSE, echo=FALSE}
@@ -162,8 +164,9 @@ Kruschke (2018) suggests that such null value could be set, by default, to the -
 
 Kruschke (2010, 2011, 2014) suggests using the proportion of the 95\% (or 90\%, considered more stable) HDI that falls within the ROPE as an index for "null-hypothesis" testing (as understood under the Bayesian framework, see [equivalence_test](https://easystats.github.io/bayestestR/reference/equivalence_test.html)).
 
-```{r message=FALSE, warning=FALSE, results='hide'}
-rope(rnorm(1000, 1, 1), range = c(-0.1, 0.1))
+```{r message=FALSE, warning=FALSE}
+posterior <- distribution_normal(100, 0.4, 0.2)
+rope(posterior, range = c(-0.1, 0.1))
 ```
 
 ```{r message=FALSE, warning=FALSE, echo=FALSE}
@@ -192,8 +195,9 @@ posterior %>%
 **Test for Practical Equivalence** based on the *"HDI+ROPE decision rule"* (Kruschke, 2018) to check whether parameter values should be accepted or rejected against an explicitly formulated "null hypothesis" (*i.e.*, a [ROPE](https://easystats.github.io/bayestestR/reference/rope.html)).
 
 
-```{r message=FALSE, warning=FALSE, results='hide'}
-equivalence_test(rnorm(1000, 1, 1), range = c(-0.1, 0.1))
+```{r message=FALSE, warning=FALSE}
+posterior <- distribution_normal(100, 0.4, 0.2)
+equivalence_test(posterior, range = c(-0.1, 0.1))
 ```
 
 
[…]
