### 2019-07-30 3796c93
plot-method for map_estimate()
