+++ NAMESPACE
@@ -101,6 +101,7 @@ S3method(plot,equivalence_test)
 S3method(plot,estimate_density)
 S3method(plot,eti)
 S3method(plot,hdi)
+S3method(plot,map_estimate)
 S3method(plot,p_direction)
 S3method(plot,point_estimate)
 S3method(plot,rope)
