### 2019-07-30 34782cf
address various reviewer comments
