+++ README.Rmd
@@ -79,7 +79,11 @@ Click on the buttons above to access the package [**documentation**](https://eas
 
 # Features
 
-[**`describe_posterior()`**](https://easystats.github.io/bayestestR/reference/describe_posterior.html) is the master function with which you can compute all of the indices cited below *at once*.
+The following figures are meant to illustrate the (statistical) concepts behind the functions. However, for most functions, `plot()`-methods are available from the [see-package](http://easystats.github.io/see).
+
+## Describing the Posterior Distribution
+
+[**`describe_posterior()`**](https://easystats.github.io/bayestestR/reference/describe_posterior.html) is the master function with which you can compute all of the indices cited below at once.
 
 ```{r message=FALSE, warning=FALSE}
 describe_posterior(rnorm(1000))
@@ -87,7 +91,6 @@ describe_posterior(rnorm(1000))
 
 
 ## Point-estimates
-
 
 ### MAP Estimate
 
@@ -110,7 +113,7 @@ posterior %>%
   geom_segment(x=0.4, xend=0.4, y=0, yend=attributes(map_result)$MAP_density, color="#E91E63", size=1) +
   geom_point(x=map_result, y=attributes(map_result)$MAP_density, color="#E91E63", size=5) +
   geom_vline(xintercept=0, linetype="dotted") +
-  theme_classic() +
+  theme_classic(base_size = 12) +
   scale_y_continuous(expand = c(0, 0), limits = c(0, 2)) +
   xlab("\nParameter Value") +
   ylab("Probability Density\n")
@@ -119,36 +122,55 @@ posterior %>%
 
 ## Uncertainty
 
-### Highest Density Interval (HDI) - *Credible* Interval (CI)
+### Highest Density Interval (HDI) and Equal-Tailed Interval (ETI)
 
-[**`hdi()`**](https://easystats.github.io/bayestestR/reference/hdi.html) computes the **Highest Density Interval (HDI)** of a posterior distribution, *i.e.*, the interval which contains all points within the interval have a higher probability density than points outside the interval. The HDI can be used in the context of Bayesian posterior characterisation as **Credible Interval (CI)**.
+[**`hdi()`**](https://easystats.github.io/bayestestR/reference/hdi.html) computes the **Highest Density Interval (HDI)** of a posterior distribution, i.e., the interval which contains all points within the interval have a higher probability density than points outside the interval. The HDI can be used in the context of Bayesian posterior characterisation as **Credible Interval (CI)**.
 
-Unlike equal-tailed intervals (see [`ci()`](https://easystats.github.io/bayestestR/reference/ci.html)) that typically exclude 2.5\% from each tail of the distribution, the HDI is *not* equal-tailed and therefore always includes the mode(s) of posterior distributions. 
+Unlike equal-tailed intervals (see [`eti()`](https://easystats.github.io/bayestestR/reference/eti.html)) that typically exclude 2.5\% from each tail of the distribution, the HDI is *not* equal-tailed and therefore always includes the mode(s) of posterior distributions. 
 
[…]
