### 2019-07-30 0824b23
add refs to readme
