+++ paper/paper.bib
@@ -366,15 +366,17 @@
 
 
 @book{kruschke2015doing,
-	doi = {10.1016/b978-0-12-405888-0.00014-3},
-	url = {https://doi.org/10.1016%2Fb978-0-12-405888-0.00014-3},
-	year = 2015,
-	publisher = {Elsevier},
-	pages = {399--416},
+	address = {Amsterdam},
+	edition = {2. ed},
+	title = {Doing {Bayesian} data analysis: a tutorial with {R}, {JAGS}, and {Stan}},
+	isbn = {978-0-12-405888-0},
+	shorttitle = {Doing {Bayesian} data analysis},
+	language = {eng},
+	publisher = {Elsevier, Academic Press},
 	author = {Kruschke, John K},
-	title = {Stan},
-	booktitle = {Doing Bayesian Data Analysis}
+	year = {2015},
 }
+
 
 
 @misc{morey_2015_blog,
