+++ README.md
@@ -84,10 +84,10 @@ cited below at once.
 
 ``` r
 describe_posterior(rnorm(1000))
-##   Parameter Median CI CI_low CI_high   pd ROPE_CI ROPE_low ROPE_high
-## 1 Posterior  0.053 89   -1.5     1.7 0.52      89     -0.1       0.1
+##   Parameter  Median CI CI_low CI_high  pd ROPE_CI ROPE_low ROPE_high
+## 1 Posterior -0.0011 89   -1.6     1.4 0.5      89     -0.1       0.1
 ##   ROPE_Percentage
-## 1           0.084
+## 1           0.095
 ```
 
 ## Point-estimates
@@ -126,12 +126,12 @@ is *not* equal-tailed and therefore always includes the mode(s) of
 posterior distributions.
 
 By default, `hdi()` returns the 89% intervals (`ci = 0.89`), deemed to
-be more stable than, for instance, 95% intervals (Kruschke, 2014). An
-effective sample size of at least 10.000 is recommended if 95% intervals
-should be computed (Kruschke 2014, p. 183ff). Moreover, 89 indicates the
-arbitrariness of interval limits - its only remarkable property is being
-the highest prime number that does not exceed the already unstable 95%
-threshold (McElreath, 2015).
+be more stable than, for instance, 95% intervals. An effective sample
+size of at least 10.000 is recommended if 95% intervals should be
+computed (Kruschke 2015). Moreover, 89 indicates the arbitrariness of
+interval limits - its only remarkable property is being the highest
+prime number that does not exceed the already unstable 95% threshold
+(McElreath 2018).
 
 ``` r
 posterior <- distribution_chisquared(100, 3)
@@ -165,20 +165,20 @@ different from 0 does not make much sense (the probability of it being
 different from a single point being infinite). Therefore, the idea
 underlining ROPE is to let the user define an area around the null value
 enclosing values that are *equivalent to the null* value for practical
-purposes (Kruschke 2010, 2011, 2014).
+purposes (Kruschke and Liddell 2018, @kruschke2018rejecting).
 
-Kruschke (2018) suggests that such null value could be set, by default,
-to the -0.1 to 0.1 range of a standardized parameter (negligible effect
-size according to Cohen, 1988). This could be generalized: For instance,
-for linear models, the ROPE could be set as `0 +/- .1 * sd(y)`. This
-ROPE range can be automatically computed for models using the
+Kruschke suggests that such null value could be set, by default, to the
+-0.1 to 0.1 range of a standardized parameter (negligible effect size
+according to Cohen, 1988). This could be generalized: For instance, for
+linear models, the ROPE could be set as `0 +/- .1 * sd(y)`. This ROPE
+range can be automatically computed for models using the
 [rope\_range](https://easystats.github.io/bayestestR/reference/rope_range.html)
 function.
 
-Kruschke (2010, 2011, 2014) suggests using the proportion of the 95% (or
-90%, considered more stable) HDI that falls within the ROPE as an index
-for “null-hypothesis” testing (as understood under the Bayesian
-framework, see
+Kruschke suggests using the proportion of the 95% (or 90%, considered
+more stable) HDI that falls within the ROPE as an index for
+“null-hypothesis” testing (as understood under the Bayesian framework,
+see
[…]
