### 2019-07-30 006b13d
ci_method for ROPE
