### 2019-07-30 a40e630
added Tristan as reviewer in DESCRIPTION (and version bump)
