+++ README.md
@@ -85,9 +85,9 @@ cited below at once.
 ``` r
 describe_posterior(rnorm(1000))
 ##   Parameter Median CI CI_low CI_high   pd ROPE_CI ROPE_low ROPE_high
-## 1 Posterior 0.0082 89   -1.5     1.6 0.51      89     -0.1       0.1
+## 1 Posterior -0.025 89   -1.7     1.5 0.51      89     -0.1       0.1
 ##   ROPE_Percentage
-## 1            0.11
+## 1             0.1
 ```
 
 ## Point-estimates
@@ -128,10 +128,10 @@ posterior distributions.
 By default, `hdi()` returns the 89% intervals (`ci = 0.89`), deemed to
 be more stable than, for instance, 95% intervals. An effective sample
 size of at least 10.000 is recommended if 95% intervals should be
-computed (Kruschke 2015). Moreover, 89 indicates the arbitrariness of
+computed (Kruschke, 2015). Moreover, 89 indicates the arbitrariness of
 interval limits - its only remarkable property is being the highest
 prime number that does not exceed the already unstable 95% threshold
-(McElreath 2018).
+(McElreath, 2018).
 
 ``` r
 posterior <- distribution_chisquared(100, 3)
@@ -165,7 +165,7 @@ different from 0 does not make much sense (the probability of it being
 different from a single point being infinite). Therefore, the idea
 underlining ROPE is to let the user define an area around the null value
 enclosing values that are *equivalent to the null* value for practical
-purposes (Kruschke and Liddell 2018, @kruschke2018rejecting).
+purposes (Kruschke & Liddell, 2018, p. @kruschke2018rejecting).
 
 Kruschke suggests that such null value could be set, by default, to the
 -0.1 to 0.1 range of a standardized parameter (negligible effect size
@@ -196,7 +196,7 @@ rope(posterior, range = c(-0.1, 0.1))
 
 [**`equivalence_test()`**](https://easystats.github.io/bayestestR/reference/equivalence_test.html)
 is a **Test for Practical Equivalence** based on the *“HDI+ROPE decision
-rule”* (Kruschke 2018) to check whether parameter values should be
+rule”* (Kruschke, 2018) to check whether parameter values should be
 accepted or rejected against an explicitly formulated “null hypothesis”
 (*i.e.*, a
 [ROPE](https://easystats.github.io/bayestestR/reference/rope.html)).
@@ -259,8 +259,8 @@ the prior and posterior odds of the parameter falling within or outside
 the null; When the null is a point, a Savage-Dickey density ratio is
 computed, which is also an approximation of a Bayes factor comparing the
 marginal likelihoods of the model against a model in which the tested
-parameter has been restricted to the point null (Wagenmakers et al.
-2010).
+parameter has been restricted to the point null (Wagenmakers, Lodewyckx,
+Kuriyal, & Grasman, 2010).
 
 ``` r
 prior <- rnorm(1000, mean = 0, sd = 1)
@@ -270,7 +270,7 @@ bayesfactor_parameters(posterior, prior, direction = "two-sided", null = 0)
 ## # Bayes Factor (Savage-Dickey density ratio)
 ## 
 ##  Bayes Factor
-##          1.71
+##          1.69
 ## 
 ## * Evidence Against The Null: [0]
 ```
@@ -356,7 +356,7 @@ Compute the density of a given point of a distribution.
 
 ``` r
[…]
