### 2019-08-01 a900bfc
smaller size + logo is now in
