### 2019-08-02 a681ba8
breaking change: default ROPE range for binomial models
