### 2019-08-02 8fa8c56
Update DESCRIPTION
R 3.6.0 is required for bf_inclusion to work. It's been 3 months since release - I think this should be fine.
