### 2019-08-02 #1541 @cMadan
Thanks for the thorough review, @tjmahr!

@DominiqueMakowski, it looks like there are no outstanding issues, is that correct?
