### 2019-08-02 ad42d31
r version error in bf_inclusion
#207
