+++ DESCRIPTION
@@ -40,7 +40,7 @@ License: GPL-3
 Encoding: UTF-8
 LazyData: true
 Depends:
-  R (>= 3.6),
+  R (>= 3.0),
   stats,
   methods,
   utils
