### 2019-08-02 97657da
Update bayesfactor_inclusion.R
#207
