### 2019-08-04 2483c85
skip BF inclusion test if R oldrel
