+++ R/bayesfactor_inclusion.R
@@ -70,7 +70,7 @@ bayesfactor_inclusion <- function(models, match_models = FALSE, prior_odds = NUL
 #' @export
 bayesfactor_inclusion.bayesfactor_models <- function(models, match_models = FALSE, prior_odds = NULL, ...) {
   if (utils::packageVersion("base") < "3.6.0") {
-    stop("'bayesfactor_inclusion' requires R 3.6.0 or later to work.\nYou might find the 'installr' package useful.")
+    stop("'bayesfactor_inclusion' requires R 3.6.0 or later to work.\nYou can easily update R using the 'installr' package.")
   }
 
 
