+++ R/bayesfactor_inclusion.R
@@ -37,11 +37,10 @@
 #' mo1 <- lm(Sepal.Length ~ Species, data = iris)
 #' mo2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
 #' mo3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
-#' \donttest{
+#' \dontrun{
 #' BFmodels <- bayesfactor_models(mo1, mo2, mo3, denominator = mo0)
 #' bayesfactor_inclusion(BFmodels)
-#' }
-#' \dontrun{
+#'
 #' # BayesFactor
 #' # -------------------------------
 #' library(BayesFactor)
