### 2019-08-04 702bf6b
hot fix for R < 3.6
#207
