+++ paper/paper.md
@@ -226,6 +226,6 @@ plot(result_pd)
 
 # Acknowledgments
 
-**bayestestR** is part of the [*easystats*](`https://github.com/easystats/easystats`) ecosystem [relying on the **insight** package to access information contained in models; @ludecke2019insight], a collaborative project created to facilitate the usage of R. Thus, we would like to thank the [council of masters](https://github.com/orgs/easystats/people) of easystats, all other padawan contributors, as well as the users.
+**bayestestR** is part of the [*easystats*](https://github.com/easystats/easystats) ecosystem [relying on the **insight** package to access information contained in models; @ludecke2019insight], a collaborative project created to facilitate the usage of R. Thus, we would like to thank the [council of masters](https://github.com/orgs/easystats/people) of easystats, all other padawan contributors, as well as the users.
 
 # References
