+++ NAMESPACE
@@ -227,5 +227,4 @@ importFrom(stats,terms.formula)
 importFrom(stats,update)
 importFrom(utils,capture.output)
 importFrom(utils,install.packages)
-importFrom(utils,packageVersion)
 importFrom(utils,stack)
