### 2019-08-05 04279e3
enabled method arg in density_at
