+++ R/p_map.R
@@ -29,7 +29,11 @@
 #' p_map(bf)
 #' }
 #'
-#' @references \href{https://www.youtube.com/watch?v=Ip8Ci5KUVRc}{Mill's talk}
+#' @seealso \href{https://www.youtube.com/watch?v=Ip8Ci5KUVRc}{Jeff Mill's talk}
+#'
+#' @references \itemize{
+#'   \item Mills, J. A. (2018). Objective Bayesian Precise Hypothesis Testing. University of Cincinnati.
+#' }
 #'
 #' @importFrom stats density
 #' @export
