### 2019-08-05 c34f798
added example for sensitivity to density and method arg in map_estimate too
