+++ man/density_at.Rd
@@ -13,7 +13,7 @@ density_at(posterior, x, precision = 2^10, method = "kernel", ...)
 
 \item{precision}{Number of points of density data. See the \code{n} parameter in \link[=density]{density}.}
 
-\item{method}{Method of density estimation. Can be \code{"kernel"} (default), \code{"logspline"} or \code{"KernSmooth"}.}
+\item{method}{Density estimation method. Can be \code{"kernel"} (default), \code{"logspline"} or \code{"KernSmooth"}.}
 
 \item{...}{Currently not used.}
 }
