+++ tests/testthat/test-bayesfactor_models.R
@@ -37,12 +37,12 @@ test_that("bayesfactor_models RSTANARM", {
   library(rstanarm)
   set.seed(444)
   junk <- capture.output(stan_bf_0 <- stan_glm(Sepal.Length ~ 1,
-                                               data = iris,
-                                               diagnostic_file = file.path(tempdir(), "df0.csv")
+    data = iris,
+    diagnostic_file = file.path(tempdir(), "df0.csv")
   ))
   junk <- capture.output(stan_bf_1 <- stan_glm(Sepal.Length ~ Species,
-                                               data = iris,
-                                               diagnostic_file = file.path(tempdir(), "df1.csv")
+    data = iris,
+    diagnostic_file = file.path(tempdir(), "df1.csv")
   ))
 
   testthat::expect_warning(bayestestR::bayesfactor_models(stan_bf_0, stan_bf_1))
