+++ R/p_map.R
@@ -34,11 +34,13 @@
 #' # Robustness to density estimation method
 #' set.seed(333)
 #' data <- data.frame()
-#' for(iteration in 1:250){
+#' for (iteration in 1:250) {
 #'   x <- rnorm(1000, 1, 1)
-#'   result <- data.frame("Kernel" = p_map(x, method = "kernel"),
-#'                        "KernSmooth" = p_map(x, method = "KernSmooth"),
-#'                        "logspline" = p_map(x, method = "logspline"))
+#'   result <- data.frame(
+#'     "Kernel" = p_map(x, method = "kernel"),
+#'     "KernSmooth" = p_map(x, method = "KernSmooth"),
+#'     "logspline" = p_map(x, method = "logspline")
+#'   )
 #'   data <- rbind(data, result)
 #' }
 #' data$KernSmooth <- data$Kernel - data$KernSmooth
