### 2019-08-06 059072c
fix bibliography typo
