+++ vignettes/bibliography.bib
@@ -1,26 +1,9 @@
-
-
-
-
-
-
-
-
-
-
-
-
-
 @book{mcelreath2018statistical,
   title={Statistical rethinking: A Bayesian course with examples in R and Stan},
   author={McElreath, Richard},
   year={2018},
   publisher={Chapman and Hall/CRC}
 }
-
-
-
-
 
 @article{wagenmakers2018bayesian,
   title={Bayesian inference for psychology. Part I: Theoretical advantages and practical ramifications},
@@ -63,8 +46,6 @@
   publisher={Nature Publishing Group}
 }
 
-
-
 @article{dienes2018four,
   title={Four reasons to prefer Bayesian analyses over significance testing},
   author={Dienes, Zoltan and Mclatchie, Neil},
@@ -76,8 +57,6 @@
   publisher={Springer}
 }
 
-
-
 @article{lakens2018equivalence,
   title={Equivalence testing for psychological research: A tutorial},
   author={Lakens, Dani{\"e}l and Scheel, Anne M and Isager, Peder M},
@@ -86,10 +65,6 @@
   year={2018},
   publisher={SAGE Publications Sage CA: Los Angeles, CA}
 }
-
-
-
-
 
 @article{etz2018bayesian,
   title={Bayesian inference and testing any hypothesis you can specify},
@@ -110,9 +85,6 @@
   year={2018},
   publisher={Springer}
 }
-
-
-
 
 @article{wagenmakers2017need,
   title={The need for Bayesian hypothesis testing in psychological science},
