### 2019-08-07 #1541 @cMadan
@DominiqueMakowski, great!!

To move forward with accepting your submission, there are a few last things to take care of:

You may find this helpful: https://guides.github.com/activities/citable-code/
