### 2019-08-07 #1541 @DominiqueMakowski
@cMadan we have tagged a [v0.2.5 release](https://github.com/easystats/bayestestR/releases/tag/v0.2.5) and a [Zenodo](https://zenodo.org/badge/latestdoi/165641861) identifier ☺️

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3361605.svg)](https://doi.org/10.5281/zenodo.3361605)
