### 2019-08-08 bfcff7a
show "plot()" in vignette
