### 2019-08-12 #1541 @cMadan
@DominiqueMakowski, perfect, thank you!

@openjournals/joss-eics, I think we're all set to accept here!
