### 2019-08-12 #1541 @danielskatz
Thanks to @paul-buerkner and @tjmahr for reviewing and @cMadan for editing
