### 2019-08-12 #1541 @danielskatz
👋 @DominiqueMakowski - please see https://github.com/easystats/bayestestR/pull/217 and merge it - also carefully check the rest of the bib to make sure I didn't miss anything else (e.g., words in lower case that should be in upper case, odd periods at the end of titles, etc.)
