### 2019-08-12 #1541 @danielskatz
In addition, here are some changes for the paper.
