+++ paper/paper.bib
@@ -450,15 +450,16 @@
 
 
 @incollection{mills2014bayesian,
+	address = {Berlin, Heidelberg},
+	title = {Bayesian {MCMC} {Estimation}},
 	doi = {10.1007/978-3-642-23430-9_89},
 	url = {https://doi.org/10.1007%2F978-3-642-23430-9_89},
-	year = 2013,
-	month = {jul},
-	publisher = {Springer Berlin Heidelberg},
-	pages = {1571--1595},
+	booktitle = {Handbook of Regional Science},
+	publisher = {Springer},
 	author = {Jeffrey A. Mills and Olivier Parent},
-	title = {Bayesian {MCMC} Estimation},
-	booktitle = {Handbook of Regional Science}
+	editor = {Manfred M. Fischer and Peter Nijkamp},
+	year = {2014},
+	pages = {1571--1595}
 }
 
 
