+++ paper/paper.bib
@@ -46,7 +46,7 @@
 	number = {1},
 	pages = {35--57},
 	author = {Eric-Jan Wagenmakers and Maarten Marsman and Tahira Jamil and Alexander Ly and Josine Verhagen and Jonathon Love and Ravi Selker and Quentin F. Gronau and Martin {\v{S}}m{\'{\i}}ra and Sacha Epskamp and Dora Matzke and Jeffrey N. Rouder and Richard D. Morey},
-	title = {Bayesian inference for psychology. Part I: Theoretical advantages and practical ramifications},
+	title = {Bayesian inference for psychology. Part {I}: Theoretical advantages and practical ramifications},
 	journal = {Psychonomic Bulletin {\&} Review}
 }
 
@@ -463,12 +463,11 @@
 }
 
 
-@article{morey2018bayesfactor,
+@manual{morey2018bayesfactor,
   title={{BayesFactor}: Computation of {B}ayes Factors for Common Designs},
   author={Richard D. Morey and Jeffrey N. Rouder},
   year={2018},
   note={R package version 0.9.12-4.2},
-  journal={[Computer software]},
   url={https://CRAN.R-project.org/package=BayesFactor},
 }
 
