### 2019-08-12 #1541 @strengejacke
Thanks for the language editing! i have gone through the references and found some minor changes. I will hand over to @DominiqueMakowski for the final check.
