### 2019-08-12 #1541 @danielskatz
Ok - please let me know when you & @DominiqueMakowski are done, then we can proceed.
