+++ paper/paper.bib
@@ -456,7 +456,7 @@
 	url = {https://doi.org/10.1007%2F978-3-642-23430-9_89},
 	booktitle = {Handbook of Regional Science},
 	publisher = {Springer},
-	author = {Jeffrey A. Mills and Olivier Parent},
+	author = {Mills, Jeffrey A and Parent, Olivier},
 	editor = {Manfred M. Fischer and Peter Nijkamp},
 	year = {2014},
 	pages = {1571--1595}
