### 2019-08-13 #1541 @DominiqueMakowski
@cMadan @danielskatz @paul-buerkner @tjmahr Thanks a lot again for your time and contributions! 😍
