### 2023-10-14 #1541 @ajstewartlang
:wave: @paul-buerkner Might you have the bandwidth to review the following submission? [[PRE REVIEW]: jointVIP: Prioritizing variables in observational study design with joint variable importance plot in R](https://github.com/ldliao/jointVIP)
